fix(wagon-wheel): CI lint + tooltip name fallback
- Move useWagonWheelQuery and hasWagonWheel out of the modal file into
  hooks/use-wagon-wheel.ts so wagon-wheel-modal.tsx only exports the
  component (react-doctor/only-export-components for fast refresh).
- Re-run prettier to match CI format check.
- Fall back to parsing the canonical " <bowler> to <batter>: ..." prefix
  from lDesc when batterName or bowlerName is null. RV players with no
  PC external_id (placeholder ids like -101/-102) aren't in
  rv_player_mapping, so the cumulative-chart wicket tooltip was showing
  "Batter out — bowler bowling" for them.
